third_party/WebKit/LayoutTests/fast/canvas/canvas-createImageBitmap-invalid-args-in-workers-expected.txt

[Worker] Test createImageBitmap with invalid arguments in workers.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Starting worker: ./resources/canvas-createImageBitmap-invalid-args-in-workers.js
PASS [Worker] createImageBitmap(null, 0, 0, 10, 10) threw exception TypeError: Failed to execute 'createImageBitmap' on 'WorkerGlobalScope': No function was found that matched the signature provided..
PASS [Worker] createImageBitmap(data, 0, 0, 10, 0) threw exception IndexSizeError: Failed to execute 'createImageBitmap' on 'WorkerGlobalScope': The source height provided is 0..
PASS [Worker] createImageBitmap(data, 0, 0, 0, 10) threw exception IndexSizeError: Failed to execute 'createImageBitmap' on 'WorkerGlobalScope': The source width provided is 0..
PASS successfullyParsed is true

TEST COMPLETE
